+++ src/app/pc/pc-fixe/pc-fixe.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { AuthService } from 'src/app/auth.service';
 import { Article } from '../../element';
 import { PcService } from '../pc.service';
 
@@ -10,15 +11,19 @@ import { PcService } from '../pc.service';
 })
 export class PcFixeComponent implements OnInit {
   articles: Array<Article> = [];
-  constructor(private service: PcService) { }
+  constructor(private service: PcService, private authService: AuthService) { }
 
   ngOnInit(): void {
     this.service.getDesktops().subscribe(computers => this.articles = computers);
   }
 
   delete(id: string): void {
-    this.service.delete(id);
-    location.reload();
+
+    const x = this.authService.estConnecte;
+    if (x()) {
+      this.service.delete(id);
+      location.reload();
+    }
   }
 
 }
