+++ src/app/pc/pc.service.ts
@@ -1,6 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
+import { AuthService } from '../auth.service';
 import { Article } from '../element';
 
 @Injectable({
@@ -9,7 +10,7 @@ import { Article } from '../element';
 export class PcService {
   private url: string = 'http://localhost:8080/tech/items/pc'
   
-  constructor(private client: HttpClient) {
+  constructor(private client: HttpClient, private authService: AuthService) {
     this.client = client
   }
 
@@ -18,6 +19,7 @@ export class PcService {
   getDesktops() : Observable<Array<Article>> { return this.client.get<Array<Article>>(`${this.url}/fixes`); }
   getAccessories() : Observable<Array<Article>> { return this.client.get<Array<Article>>(`${this.url}/accessoires`); }
   delete(id: string): void { 
-    this.client.delete(`http://localhost:8080/tech/item/${id}`).subscribe(() => console.log('working')); 
+    const x = this.authService.estConnecte;
+    if (x()) this.client.delete(`http://localhost:8080/tech/item/${id}`).subscribe(() => console.log('working')); 
   }
 }
