+++ src/app/stockage/cle-usb/cle-usb.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { AuthService } from 'src/app/auth.service';
 import { Article } from 'src/app/element';
 import { StockageService } from '../stockage.service';
 
@@ -10,15 +11,18 @@ import { StockageService } from '../stockage.service';
 })
 export class CleUsbComponent implements OnInit {
   articles: Article[] = [];
-  constructor(private service: StockageService) { }
+  constructor(private service: StockageService, private authService: AuthService) { }
 
   ngOnInit(): void {
     this.service.getUsb().subscribe(usbs => this.articles = usbs);
   }
 
-  delete(id: string): void {
-    this.service.delete(id);
-    location.reload();
+  delete(id: string): void { 
+    const x = this.authService.estConnecte;
+    if (x()) {
+      this.service.delete(id);
+      location.reload();
+    }
   }
 
 }
