+++ src/app/stockage/stockage-accessoire/stockage-accessoire.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { AuthService } from 'src/app/auth.service';
 import { Article } from 'src/app/element';
 import { StockageService } from '../stockage.service';
 
@@ -11,14 +12,17 @@ import { StockageService } from '../stockage.service';
 export class StockageAccessoireComponent implements OnInit {
   articles: Array<Article> = [];
 
-  constructor(private service: StockageService) { }
+  constructor(private service: StockageService, private authService: AuthService) { }
 
   ngOnInit(): void {
     this.service.getAccessories().subscribe(items => this.articles = items);
   }
 
   delete(id: string): void {
-    this.service.delete(id);
-    location.reload();
+    const x = this.authService.estConnecte;
+    if (x()) {
+      this.service.delete(id);
+      location.reload();
+    }
   }
 }
