+++ src/app/telephone/telephone-list/telephone-list.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { AuthService } from 'src/app/auth.service';
 import { Article } from 'src/app/element';
 import { TelephoneService } from '../telephone.service';
 
@@ -10,14 +11,17 @@ import { TelephoneService } from '../telephone.service';
 })
 export class TelephoneListComponent implements OnInit {
   articles: Array<Article> = [];
-  constructor(private service: TelephoneService) { }
+  constructor(private service: TelephoneService, private authService: AuthService) { }
 
   ngOnInit(): void {
     this.service.getPhones().subscribe(phones => this.articles = phones);
   }
 
   delete(id: string): void {
-    this.service.delete(id);
-    location.reload();
+    const x = this.authService.estConnecte;
+    if (x()) {
+      this.service.delete(id);
+      location.reload();
+    }
   }
 }
